requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
allure-pytest=2.10.0=pypi_0
allure-python-commons=2.10.0=pypi_0
anyio=3.6.1=pypi_0
assertpy=1.1=pypi_0
async-generator=1.10=pypi_0
attrs=22.1.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
brotli=1.0.9=pypi_0
bzip2=1.0.8=h1de35cc_0
ca-certificates=2022.07.19=hecd8cb5_0
certifi=2022.6.15=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
configargparse=1.5.3=pypi_0
contourpy=1.0.5=pypi_0
curlify=2.2.1=pypi_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
execnet=1.9.0=pypi_0
exrex=0.10.5=pypi_0
faker=14.2.0=pypi_0
fast-map=0.0.7=pypi_0
fastapi=0.85.0=pypi_0
fastapi-utils=0.2.1=pypi_0
flask=2.2.2=pypi_0
flask-basicauth=0.2.0=pypi_0
flask-cors=3.0.10=pypi_0
fonttools=4.37.2=pypi_0
gevent=21.12.0=pypi_0
geventhttpclient=2.0.2=pypi_0
greenlet=1.1.2=pypi_0
gunicorn=20.1.0=pypi_0
h11=0.13.0=pypi_0
idna=3.4=pypi_0
influxdb-client=1.32.0=pypi_0
iniconfig=1.1.1=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
jsonpath-ng=1.5.3=pypi_0
kiwisolver=1.4.4=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.3=hb1e8313_2
locust=2.12.1=pypi_0
lxml=4.9.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.0=pypi_0
msgpack=1.0.4=pypi_0
ncurses=6.3=hca72f7f_3
numpy=1.23.3=pypi_0
openssl=1.1.1q=hca72f7f_0
outcome=1.2.0=pypi_0
packaging=21.3=pypi_0
pandas=1.4.4=pypi_0
pillow=9.2.0=pypi_0
pip=22.1.2=pypi_0
pluggy=1.0.0=pypi_0
ply=3.11=pypi_0
prometheus-client=0.14.1=pypi_0
psutil=5.9.2=pypi_0
psycogreen=1.0.2=pypi_0
psycopg2-binary=2.9.3=pypi_0
py=1.11.0=pypi_0
pydantic=1.10.2=pypi_0
pyjwt=2.5.0=pypi_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pypi_0
pytest=7.1.3=pypi_0
pytest-forked=1.4.0=pypi_0
pytest-json-report=1.5.0=pypi_0
pytest-metadata=2.0.2=pypi_0
pytest-order=1.0.1=pypi_0
pytest-rerunfailures=10.2=pypi_0
pytest-xdist=2.5.0=pypi_0
python=3.10.4=hdfd78df_0
python-dateutil=2.8.2=pypi_0
python-dotenv=0.21.0=pypi_0
pytz=2022.2.1=pypi_0
pyzmq=24.0.0=pypi_0
reactivex=4.0.4=pypi_0
readline=8.1.2=hca72f7f_1
requests=2.28.1=pypi_0
roundrobin=0.0.4=pypi_0
selenium=4.4.3=pypi_0
setuptools=63.4.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlalchemy=1.4.41=pypi_0
sqlite=3.39.2=h707629a_0
starlette=0.20.4=pypi_0
tk=8.6.12=h5d9f67b_0
tomli=2.0.1=pypi_0
torch=1.12.1=pypi_0
torchvision=0.13.1=pypi_0
tqdm=4.64.1=pypi_0
trio=0.21.0=pypi_0
trio-websocket=0.9.2=pypi_0
typing-extensions=4.3.0=pypi_0
tzdata=2022c=h04d1e81_0
unitest=1.4.27=pypi_0
urllib3=1.26.12=pypi_0
uvicorn=0.18.3=pypi_0
webdriver-manager=3.8.3=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wsproto=1.2.0=pypi_0
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_3
zope-event=4.5.0=pypi_0
zope-interface=5.4.0=pypi_0
